+++ spine-unity/Assets/Spine/Editor/spine-unity/Editor/Utility/SpineHandles.cs
@@ -560,6 +560,48 @@ namespace Spine.Unity.Editor {
 			Handles.color = savedColor;
 		}
 
+		public static void DrawPivotHandle (SkeletonGraphic skeletonGraphic, Color color) {
+			// Note: not limiting to current.type == EventType.Repaint because the FreeMoveHandle requires interaction.
+
+			float handleSize = HandleUtility.GetHandleSize(skeletonGraphic.transform.position);
+			float controlSize = handleSize * 0.15f;
+			float discSize = handleSize * 0.05f;
+			Vector3 snap = Vector3.zero;
+			Color savedColor = Handles.color;
+
+			Handles.color = color;
+			Vector3 worldSpacePivot = skeletonGraphic.rectTransform.position;
+			EditorGUI.BeginChangeCheck();
+
+#if FREE_MOVE_HANDLE_HAS_NO_ROTATION_PARAM
+			Vector3 newWorldSpacePosition = Handles.FreeMoveHandle(worldSpacePivot, controlSize, snap, Handles.CircleHandleCap);
+#else
+			Vector3 newWorldSpacePosition = Handles.FreeMoveHandle(worldSpacePivot, Quaternion.identity, controlSize, snap, Handles.CircleHandleCap);
+#endif
+			if (EditorGUI.EndChangeCheck()) {
+				RectTransform rectTransform = skeletonGraphic.rectTransform;
+				Undo.RecordObjects(new UnityEngine.Object[] { skeletonGraphic, rectTransform }, "Change Pivot");
+				Vector3 localPivot = rectTransform.InverseTransformPoint(newWorldSpacePosition);
+
+				Rect rect = rectTransform.rect;
+				Vector2 oldPivot = rectTransform.pivot;
+				Vector2 clampedPivot = Rect.PointToNormalized(rect, localPivot);
+				Vector2 localMovement = Vector2.Scale(clampedPivot - oldPivot, rect.size);
+				rectTransform.pivot = clampedPivot;
+				rectTransform.position = worldSpacePivot + rectTransform.TransformVector(localMovement);
+				newWorldSpacePosition = rectTransform.position;
+
+				Vector2 scaledOffset = skeletonGraphic.GetScaledPivotOffset();
+				scaledOffset -= localMovement;
+				skeletonGraphic.SetScaledPivotOffset(scaledOffset);
+
+				skeletonGraphic.UpdateBuffersToInstructions(true);
+				skeletonGraphic.UpdateMeshAndMaterialsToBuffers();
+			}
+			Handles.DrawSolidDisc(newWorldSpacePosition, skeletonGraphic.transform.forward, discSize);
+			Handles.color = savedColor;
+		}
+
 		static void DrawCrosshairs2D (Vector3 position, float scale, float skeletonRenderScale = 1f) {
 			if (UnityEngine.Event.current.type != EventType.Repaint) return;